Created datamodel.applicationsolution-add-attribute.xml
This XML file was created for the extension to modify iTop's status fields in order to add additional options. Before upgrading iTop, the XML file must contain no modifications to the XML structure; after the upgrading process runs successfully and the extension is added in the wizard, the XML file is modified to reflect the desired changes to iTop's fields.